requirements.txt
asgiref==3.4.1
attrs==20.3.0
Automat==20.2.0
boto3==1.17.47
botocore==1.20.47
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==3.4.7
cssselect==1.1.0
fastapi==0.65.3
fpdf==1.7.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httptools==0.2.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
parsel==1.6.0
pdfrw==0.4
Pillow==8.2.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-dateutil==2.8.1
python-dotenv==0.17.0
PyYAML==5.4.1
queuelib==1.5.0
requests==2.25.1
s3transfer==0.3.6
Scrapy==2.5.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
starlette==0.14.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typer==0.4.0
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.14.0
w3lib==1.22.0
watchgod==0.7
webdriver-manager==3.3.0
websockets==9.1
zope.interface==5.3.0
